Add second patch to roll back ffmpeg
Chromium has moved to using FFmpeg 5.1, which Arch Linux does not have
yet.

The first patch is still needed. It seems odd, given that the patch was
made in April, and they were using a bundled version of 4.4.git at the
time. Did some of the changes not make it into 5.0?